Feature/login

## 📝 Descrição
Este Pull Request implementa a funcionalidade de **Login/Autenticação** de usuários no sistema da Associação Gaia. Foram criados os endpoints, regras de serviço, validações e tratamentos de exceção necessários para garantir que o fluxo de autenticação funcione corretamente.

## 🛠️ O que foi feito?
- **Controllers & DTOs:** Criação do `LoginDTO` para capturar as credenciais e atualização da `UsuarioController` para gerenciar as requisições de login.
- **Service & Repositories:** Implementação da lógica de validação de credenciais no `UsuarioService` integrado aos repositórios.
- **Tratamento de Erros:** Criação da exceção personalizada `CredenciaisInvalidasException` e mapeamento global no `GlobalExceptionHandler` para retornar respostas limpas e semânticas ao cliente.
- **Modelos afetados:** Ajustes pontuais nos modelos `Usuario`, `Artesao` e `Trabalho`.
- **Testes:** Criação e atualização de testes unitários para a controller e a service de usuários.

## 🧪 Como testar?
1. Certifique-se de aplicar as configurações necessárias de banco no `application.properties`.
2. Envie uma requisição `POST` para o endpoint `/api/usuarios/login` (ou o endpoint de login correspondente) contendo o JSON com as credenciais.
3. Valide o comportamento de sucesso e o de falha (com credenciais incorretas) para ver o tratamento de erro em ação.

+++ src/main/java/br/edu/ifsp/associacaogaia/model/Artesao.java
@@ -12,7 +12,7 @@ public class Artesao {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_artesao")
-    private Integer idArtesao;
+    private Long idArtesao;
 
     @OneToOne
     @JoinColumn(name = "id_usuario", nullable = false, unique = true)
@@ -63,7 +63,7 @@ public class Artesao {
     protected Artesao() {
     }
 
-    public Integer getIdArtesao() {
+    public Long getIdArtesao() {
         return idArtesao;
     }
