+++ src/main/java/br/edu/ifsp/associacaogaia/repository/TrabalhoRepository.java
@@ -3,5 +3,5 @@ package br.edu.ifsp.associacaogaia.repository;
 import br.edu.ifsp.associacaogaia.model.Trabalho;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrabalhoRepository extends JpaRepository<Trabalho, Integer> {
+public interface TrabalhoRepository extends JpaRepository<Trabalho, Long> {
 }
